Experimenter

Documentation for Experimenter. Head over to Getting Started to get an overview of this package.

Experimenter.Experiment

Experimenter.count_values

Experimenter.extract_value

Experimenter.merge_databases!

Experimenter.Experiment — Type

Experiment

A database object for storing the configuration options of an experiment.

The signature of the function supplied should be:

fn(configuration::Dict{Symbol, Any}, trial_id::UUID)

The function should be available when including the file provided.

A name is required to uniquely label this experiment.

Experimenter.count_values — Method

count_values(entry)

Counts the number of different values which 'variable' can take.

Experimenter.extract_value — Method

extract_value(variable, i)

Returns the i^th possible value of 'variable'. 'i' follows 1 based indexing.

Experimenter.merge_databases! — Method

merge_databases!(primary_db, secondary_db)

Searches all of the records from the secondary database and adds them to the first database.